+++ app/src/main/java/team_h/boostcamp/myapplication/data/repository/DiaryRepositoryContract.java
@@ -5,13 +5,13 @@ import java.util.List;
 import androidx.annotation.NonNull;
 import io.reactivex.Completable;
 import io.reactivex.Single;
-import team_h.boostcamp.myapplication.data.remote.deepaffects.request.EmotionAnalyzeRequest;
+import team_h.boostcamp.myapplication.data.remote.apis.deepaffects.request.EmotionAnalyzeRequest;
 import team_h.boostcamp.myapplication.model.Diary;
 
 public interface DiaryRepositoryContract {
 
     @NonNull
-    Single<List<Diary>> loadMoreDiaryItems(final int idx);
+    Single<List<Diary>> loadMoreDiaryItems(final long idx);
 
     @NonNull
     Single<Integer> analyzeVoiceEmotion(@NonNull EmotionAnalyzeRequest request);
